+++ backend/config/bootstrap.php
@@ -1,4 +1,8 @@
 <?php
 /**
+<<<<<<< HEAD
  * @author Andrey Avseenko <bropwnz0r@gmail.com>
+=======
+ * @author Eugene Terentev <eugene@terentev.net>
+>>>>>>> upstream/master
  */