+++ backend/controllers/CacheController.php
@@ -1,6 +1,10 @@
 <?php
 /**
+<<<<<<< HEAD
  * Avseenko Andrey <bropwnz0r@gmail.com>
+=======
+ * Eugine Terentev <eugine@terentev.net>
+>>>>>>> upstream/master
  */
 
 namespace backend\controllers;