+++ common/behaviors/LocaleBehavior.php
@@ -4,6 +4,7 @@ namespace common\behaviors;
 
 use yii\base\Behavior;
 use Yii;
+use yii\web\Application;
 
 /**
  * Class LocaleBehavior
@@ -17,12 +18,16 @@ class LocaleBehavior extends Behavior
     public $cookieName = '_locale';
 
     /**
+     * @var bool
+     */
+    public $enablePreferredLanguage = true;
+    /**
      * @return array
      */
     public function events()
     {
         return [
-            \yii\web\Application::EVENT_BEFORE_REQUEST => 'beforeRequest'
+            Application::EVENT_BEFORE_REQUEST => 'beforeRequest'
         ];
     }
 
@@ -37,12 +42,21 @@ class LocaleBehavior extends Behavior
         ) {
             $userLocale = Yii::$app->getRequest()->getCookies()->getValue($this->cookieName);
         } else {
-            $userLocale = !Yii::$app->user->isGuest
-                        && Yii::$app->user->getIdentity()->userProfile
-                        && Yii::$app->user->getIdentity()->userProfile->locale
-                ? Yii::$app->user->getIdentity()->userProfile->locale
-                : Yii::$app->request->getPreferredLanguage(array_keys(Yii::$app->params['availableLocales']));
+            $userLocale = Yii::$app->language;
+            if (!Yii::$app->user->isGuest && Yii::$app->user->identity->userProfile->locale) {
+                $userLocale = Yii::$app->user->getIdentity()->userProfile->locale;
+            } elseif ($this->enablePreferredLanguage) {
+                $userLocale = Yii::$app->request->getPreferredLanguage($this->getAvailableLocales());
+            }
         }
         Yii::$app->language = $userLocale;
+    }
+
+    /**
+     * @return array
+     */
+    protected function getAvailableLocales()
+    {
+        return array_keys(Yii::$app->params['availableLocales']);
     }
 }